// frontend/src/services/interfaces/emailLog.ts
// Generated by https://quicktype.io

export interface EmailLogResponse {
  statusCode: number
  data: EmailLog[]
}

export interface EmailLog {
  id: number
  sender: string
  recipient: string
  subject: string
  body: string
  provider: string
  errorMessage: null | string
  isError: boolean
  createdAt: string
}

export enum StatusTypes {
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}

export interface TransformEmailLog extends EmailLog {
  _raw: EmailLog
}

export type TransformEmailLogs = Record<StatusTypes, TransformEmailLog[]>

export function getIntialEmailLogs(): TransformEmailLogs {
  return {
    [StatusTypes.SUCCESS]: [],
    [StatusTypes.FAILED]: [],
  }
}